Refactor todoList
