+++ src/js/app.js
@@ -2,6 +2,7 @@ import { removeRouter, setRouter } from './routes/router.js';
 import { setEvent, removeEvent } from './utils/event.js';
 import { renderView } from './views/todoList.js';
 
+
 export default () => {
   return {
     init: () => {
