+++ src/js/routes/router.js
@@ -1,19 +1,8 @@
+import { filterBy } from "../store/todoList.js";
 
-const filter = {
-  active: () => filteredList = todoList.filter(item => !item.completed),
-  completed: () => filteredList = todoList.filter(item => item.completed),
-  all: () => filteredList = [],
-}
-
-const filterBy = ({ newURL }) => {
+export const setRouter = () => window.addEventListener('hashchange', ({ newURL }) => {
   const type = newURL.split('#')[1];
-  const filterType = filter[type];
-  if (!filterType) return;
-
-  filterType();
-  renderView();
-}
-
-export const setRouter = () => window.addEventListener('hashchange', filterBy);
+  filterBy(type);
+});
 
 export const removeRouter = () =>   window.removeEventListener('hashchange', filterBy);
