+++ src/js/utils/element.js
@@ -1,11 +1,13 @@
-export const getEletemt = (className) => document.querySelector(className);
+import { eventKeys } from "./event.js";
+
+export const getElement = (className) => document.querySelector(className);
 
 export const addClassName = ($element, className) => $element.classList.add(className);
 
 export const removeClassName = ($element, className) => $element.classList.remove(className);
 
 export const getTriggerEventName = (event) => {
-  if (keys.includes(event.key)) return event.key;
+  if (eventKeys.includes(event.key)) return event.key;
   return event.target.dataset.event;
 }
