+++ src/js/utils/event.js
@@ -1,11 +1,12 @@
-import { getTriggerEventName } from './element.js';
-import { upsertItem, deleteItem, stopEditing } from './todoList.js';
-import { getEletemt } from './todoListRenderer.js';
+import { upsertItem, deleteItem } from "../store/todoList.js";
+import { getEletemt, getTriggerEventName } from "./element.js";
+import { stopEditing } from "./helper.js";
 
 const $todoapp = getEletemt('.todoapp');
 const $todoList = getEletemt('.todo-list');
 
 const eventTypes = ['keyup', 'click', 'change', 'dblclick'];
+export const eventKeys = ['Enter', 'Escape'];
 
 const event = {
   onkeyup: {
@@ -42,5 +43,8 @@ export const setEvent = () => {
 }
 
 export const removeEvent = () => {
-  eventTypes.forEach(type => $todoList.removeEventListener(type, eventHandler));
+  eventTypes.forEach(type => {
+    const $element = type === 'keyup' ? $todoapp : $todoList;
+    $element.removeEventListener(type, eventHandler)
+  });
 }
